Refactor into AlertTable
